refactor: rooms
