--- candle/entities/room/templates/room/list_rooms.html
@@ -1,22 +1,0 @@
-{# Print out all rooms #}
-
-{% extends "main/base.html" %}
-
-{% block obsah_in %}
-    {{ super() }}
-        <p>
-            {% for room_prefix in rooms_dict %}
-            <a href="#{{ room_prefix }}">{{ room_prefix }}</a>
-            {% endfor %}
-        </p>
-
-        {# For each categeroy:#}
-        {% for room_prefix in rooms_dict %}
-        <h2 id="{{ room_prefix }}">{{ room_prefix }}</h2>
-            <ul>
-                {% for room in rooms_dict[room_prefix] -%}
-                    <li><a href="{{ url_for('room.show_timetable', slug=room.url_id) }}">{{ room.name }}</a></li>
-                {%- endfor %}
-            </ul>
-        {% endfor %}
-{% endblock %}
