+++ candle/panel/templates/panel/panel.html
@@ -29,7 +29,7 @@
         <div class="panel_cast_obsah">
                 {% include('search/room_search.html') %}
             <div class="other_links">
-                <a href="{{ url_for('room.list_rooms') }}">Rozvrhy všetkých miestností</a>
+                <a href="{{ url_for('rooms.listing') }}">Rozvrhy všetkých miestností</a>
             </div>
         </div>
     </div>
