+++ candle/rooms/templates/rooms/listing.html
@@ -1,0 +1,22 @@
+{# Print out all rooms #}
+
+{% extends "main/base.html" %}
+
+{% block obsah_in %}
+    {{ super() }}
+
+    <p>
+        {% for room_prefix in rooms_dict %}
+        <a href="#{{ room_prefix }}">{{ room_prefix }}</a>
+        {% endfor %}
+    </p>
+
+    {% for room_prefix in rooms_dict %}
+    <h2 id="{{ room_prefix }}">{{ room_prefix }}</h2>
+        <ul>
+            {% for room in rooms_dict[room_prefix] -%}
+                <li><a href="{{ url_for('rooms.show_timetable', slug=room.url_id) }}">{{ room.name }}</a></li>
+            {%- endfor %}
+        </ul>
+    {% endfor %}
+{% endblock %}
